workbench-ai/requirements.txt
fastapi
uvicorn
pinecone
openai
python-dotenv
pydantic
